feat: add exceptions
