+++ src/main/java/nextstep/common/exception/member/AuthenticationException.java
@@ -1,0 +1,31 @@
+package nextstep.common.exception.member;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class AuthenticationException extends RuntimeException {
+
+	private HttpStatus status = HttpStatus.UNAUTHORIZED;
+
+	public AuthenticationException() {
+		super();
+	}
+
+	public AuthenticationException(String message) {
+		super(message);
+	}
+
+	public AuthenticationException(HttpStatus status) {
+		super(status.getReasonPhrase());
+	}
+
+	public AuthenticationException(String message, HttpStatus status) {
+		super(message);
+		this.status = status;
+	}
+
+	public HttpStatus getStatus() {
+		return status;
+	}
+}
