+++ src/main/java/nextstep/common/exception/member/AuthorizationException.java
@@ -1,0 +1,31 @@
+package nextstep.common.exception.member;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class AuthorizationException extends RuntimeException {
+
+	private HttpStatus status = HttpStatus.UNAUTHORIZED;
+
+	public AuthorizationException() {
+		super();
+	}
+
+	public AuthorizationException(String message) {
+		super(message);
+	}
+
+	public AuthorizationException(HttpStatus status) {
+		super(status.getReasonPhrase());
+	}
+
+	public AuthorizationException(String message, HttpStatus status) {
+		super(message);
+		this.status = status;
+	}
+
+	public HttpStatus getStatus() {
+		return status;
+	}
+}
