+++ src/main/java/nextstep/common/exception/subway/SectionCreationNotValidException.java
@@ -1,0 +1,34 @@
+package nextstep.common.exception.subway;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+/**
+ * @author : Rene Choi
+ * @since : 2024/01/31
+ */
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class SectionCreationNotValidException extends RuntimeException {
+	private HttpStatus status = HttpStatus.BAD_REQUEST;
+
+	public SectionCreationNotValidException() {
+		super();
+	}
+
+	public SectionCreationNotValidException(String message) {
+		super(message);
+	}
+
+	public SectionCreationNotValidException(HttpStatus status) {
+		super(status.getReasonPhrase());
+	}
+
+	public SectionCreationNotValidException(String message, HttpStatus status) {
+		super(message);
+		this.status = status;
+	}
+
+	public HttpStatus getStatus() {
+		return status;
+	}
+}
