+++ src/main/java/nextstep/common/exception/subway/SectionDeletionNotValidException.java
@@ -1,4 +1,4 @@
-package nextstep.common.exception;
+package nextstep.common.exception.subway;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -8,22 +8,22 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  * @since : 2024/01/31
  */
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class SectionCreationNotValidException extends RuntimeException {
+public class SectionDeletionNotValidException extends RuntimeException {
 	private HttpStatus status = HttpStatus.BAD_REQUEST;
 
-	public SectionCreationNotValidException() {
+	public SectionDeletionNotValidException() {
 		super();
 	}
 
-	public SectionCreationNotValidException(String message) {
+	public SectionDeletionNotValidException(String message) {
 		super(message);
 	}
 
-	public SectionCreationNotValidException(HttpStatus status) {
+	public SectionDeletionNotValidException(HttpStatus status) {
 		super(status.getReasonPhrase());
 	}
 
-	public SectionCreationNotValidException(String message, HttpStatus status) {
+	public SectionDeletionNotValidException(String message, HttpStatus status) {
 		super(message);
 		this.status = status;
 	}
